notification

+++ SRSN/SRSN.DatabaseManager/Entities/UserReactionRecipe.cs
@@ -8,7 +8,7 @@ namespace SRSN.DatabaseManager.Entities
         public int UserId { get; set; }
         public int RecipeId { get; set; }
         public int? TotalView { get; set; }
-        public int? RatingRecipe { get; set; }
+        public double? RatingRecipe { get; set; }
         public double? Point { get; set; }
         public bool? IsView { get; set; }
         public bool? IsLike { get; set; }
